// src/utils/types.ts
export interface QuizData {
  symptoms?: string[];
  objectives?: string[];
  dietaryHabits?: string[];
  lifestyle?: {
    sleep?: number;
    stress?: string;
    exercise?: string;
    sleepQuality?: string;
    lifestyleFactors?: string[];
  };
  proteinConsumption?: string;
  userInfo?: {
    age?: string | number;
    gender?: string;
    email?: string;
  };
  medications?: string[];
}